safari-534.53-branch/LayoutTests/fast/dom/HTMLFormElement/move-option-between-documents-expected.txt

PASS typeof select.options.namedItem('option') is 'undefined'
PASS successfullyParsed is true

TEST COMPLETE
